+++ src/lib.rs
@@ -7,6 +7,7 @@ pub mod fence;
 pub mod freeze;
 pub mod include;
 pub mod install;
+pub mod list_of_listings;
 pub mod manifest;
 pub mod number;
 pub mod verify;